undo

+++ src/main/java/Commands.java
@@ -34,24 +34,7 @@ public class Commands extends ListenerAdapter {
 
         System.out.println("Messaged received from " + event.getAuthor().getName() + ": " +event.getMessage().getContentDisplay());
 
-        int commandNum = 0; //number to convernt msg to command num
-        if (msg.equals("!cf") || msg.equals("!coinflip")) {
-            coinflip();
-        } else if (msg.startsWith("!repeat")) {
-            repeat();
-        } else if (msg.startsWith("!whoami")) {
-            whoAmI();
-        } else if (msg.startsWith("!pfp") || msg.startsWith("!profilepic")) {
-            profilePic();
-        } else if (msg.equals("!ping")) {
-            onlineStatus();
-        } else if (msg.startsWith("!status")) {
-            statusMessage();
-        } else if (msg.equals("!shutdown")) {
-            shutdown();
-        }
-
-        private void coinflip() {
+        if (msg.equals("!cf") || msg.equals("!coinflip")) { //Coinflip
             Random ran = new Random();
             int num = ran.nextInt(100);
             if (num < 50) {
@@ -61,22 +44,18 @@ public class Commands extends ListenerAdapter {
             } else {
                 channel.sendMessage("Perfect even! Reroll").queue();
             }
-        }
 
-        private void repeat() {
+        } else if (msg.startsWith("!repeat")) {// Repeat command
             String[] temp = msg.split(" ", 2);
             channel.sendMessage(temp[1]).queue();
             message.delete().queue();
-        }
 
-        private void whoAmI() {
+        } else if (msg.startsWith("!whoami")) { //Who am i command
             channel.sendMessage("You are " + author + "!").queue();
             channel.sendMessage("Your ID is " + author.getId()).queue();
             channel.sendMessage("Lets see if you're a bot: " + author.isBot()).queue();
-        } 
-        
-        
-        private void profilePic() {
+
+        } else if (msg.startsWith("!pfp") || msg.startsWith("!profilepic")) { //Profile pic command
             if (msg.startsWith("!pfp") && msg.length() > 4 || msg.startsWith("!profilepic") && msg.length() > 11) {
                 List<User> mention = message.getMentionedUsers();
                 for (int i = 0; i < mention.size(); i++) { //Can now mention multiple people and get all avatars
@@ -95,28 +74,26 @@ public class Commands extends ListenerAdapter {
                 }
                 channel.sendMessage(author.getAvatarUrl()).queue();
             }
-        } 
-        
-        private void onlineStatus() {
-            channel.sendMessage("Pong! I'm online!").queue();
-        } 
 
-        private void statusMessage() {
+        } else if (msg.equals("!ping")) { //Ping command to check if online
+            channel.sendMessage("Pong! I'm online!").queue();
+
+        } else if (msg.startsWith("!status")) { //Set status command
             if (author.getId().equals("286647094456877056")) {
                 String[] temp = msg.split(" ", 2);
                 bot.getPresence().setActivity(Activity.playing(temp[1]));
             } else {
                 channel.sendMessage("You are not the owner, this command doesn't work for you").queue();
             }
-        } 
 
-        private void shutdown() {
+        } else if (msg.equals("!shutdown")) { //Shutdown command
             if (author.getId().equals("286647094456877056")) {
                 channel.sendMessage("Shutting down, goodbye!").queue();
                 bot.shutdown();
             } else {
                 channel.sendMessage("You are not the owner, this command doesn't work for you").queue();
             }
+
         }
 
     }
